+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [2.4.0] - 25-03-2017
 ### Added
 - Added `minRSAKeySize` helper to calculate minimum RSA key size for specified data size
 - Added zlib compression and decompression functions